Starting to fix hunter pets and stables for them. Needs more testing to get it fully working. Plus i removed the ugly checks i had to know where i was in the code.

+++ src/server/game/Server/WorldSession.h
@@ -603,7 +603,7 @@ class WorldSession
         void HandleBuyStableSlot(WorldPacket& recvPacket);
         void HandleStableRevivePet(WorldPacket& recvPacket);
         void HandleStableSwapPet(WorldPacket& recvPacket);
-        void HandleStableSwapPetCallback(PreparedQueryResult result, uint32 petId);
+        void HandleStableSwapPetCallback(PreparedQueryResult result, uint8 petId);
         void HandleStableSwapPet_(WorldPacket& recvPacket);
         void SendTrainerBuyFailed(uint64 guid, uint32 spellId, uint32 reason);
